리펙토링

+++ core/src/main/kotlin/com/bindglam/felis/utils/math/RGBAColor.kt
@@ -1,17 +1,17 @@
 package com.bindglam.felis.utils.math
 
 interface RGBAColor : RGBColor {
-    fun a(): Float
+    fun a(): Short
 
     companion object {
         @JvmStatic
-        fun of(r: Float, g: Float, b: Float, a: Float): RGBAColor = Impl(r, g, b, a)
+        fun of(r: Short, g: Short, b: Short, a: Short): RGBAColor = Impl(r, g, b, a)
     }
 
-    private class Impl(private val r: Float, private val g: Float, private val b: Float, private val a: Float) : RGBAColor {
-        override fun r(): Float = r
-        override fun g(): Float = g
-        override fun b(): Float = b
-        override fun a(): Float = a
+    private class Impl(private val r: Short, private val g: Short, private val b: Short, private val a: Short) : RGBAColor {
+        override fun r(): Short = r
+        override fun g(): Short = g
+        override fun b(): Short = b
+        override fun a(): Short = a
     }
 }
